# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"BakersChoice API","version":"0.1.0"},"paths":{"/api/auth/register":{"post":{"tags":["auth"],"summary":"Register User","operationId":"register_user_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"tags":["auth"],"summary":"Login For Access Token","operationId":"login_for_access_token_api_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_api_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["auth"],"summary":"Read Users Me","operationId":"read_users_me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/products/categories":{"get":{"tags":["products"],"summary":"Get Categories","operationId":"get_categories_api_products_categories_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Category"},"type":"array","title":"Response Get Categories Api Products Categories Get"}}}}}},"post":{"tags":["products"],"summary":"Create Category","operationId":"create_category_api_products_categories_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/products/":{"get":{"tags":["products"],"summary":"Get Products","operationId":"get_products_api_products__get","parameters":[{"name":"category_id","in":"query","required":false,"schema":{"type":"string","format":"uuid","title":"Category Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Item"},"title":"Response Get Products Api Products  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["products"],"summary":"Create Product","operationId":"create_product_api_products__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Item"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/products/{product_id}":{"put":{"tags":["products"],"summary":"Update Product","operationId":"update_product_api_products__product_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Product Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Item"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["products"],"summary":"Delete Product","operationId":"delete_product_api_products__product_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Product Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/products/{product_id}/stock":{"put":{"tags":["products"],"summary":"Update Stock","operationId":"update_stock_api_products__product_id__stock_put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Product Id"}},{"name":"change_amount","in":"query","required":true,"schema":{"type":"integer","title":"Change Amount"}},{"name":"reason","in":"query","required":true,"schema":{"$ref":"#/components/schemas/AuditReasonEnum"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Item"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/orders/":{"get":{"tags":["orders"],"summary":"Get Orders","operationId":"get_orders_api_orders__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Order"},"type":"array","title":"Response Get Orders Api Orders  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["orders"],"summary":"Create Order","operationId":"create_order_api_orders__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/orders/my-orders":{"get":{"tags":["orders"],"summary":"Get My Orders","operationId":"get_my_orders_api_orders_my_orders_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Order"},"type":"array","title":"Response Get My Orders Api Orders My Orders Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/orders/{order_id}/status":{"patch":{"tags":["orders"],"summary":"Update Order Status","operationId":"update_order_status_api_orders__order_id__status_patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Order Id"}},{"name":"status","in":"query","required":true,"schema":{"$ref":"#/components/schemas/OrderStatusEnum"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AuditReasonEnum":{"type":"string","enum":["order","restock","manual_correction","cancellation"],"title":"AuditReasonEnum"},"Body_login_for_access_token_api_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_api_auth_login_post"},"Category":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"parent_category_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Parent Category Id"},"id":{"type":"string","format":"uuid","title":"Id"}},"type":"object","required":["name","id"],"title":"Category"},"CategoryCreate":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"parent_category_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Parent Category Id"}},"type":"object","required":["name"],"title":"CategoryCreate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Item":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"price":{"type":"number","title":"Price"},"category_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Category Id"},"image_urls":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Image Urls"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags"},"stock_quantity":{"type":"integer","title":"Stock Quantity","default":0},"low_stock_threshold":{"type":"integer","title":"Low Stock Threshold","default":10},"status":{"$ref":"#/components/schemas/ItemStatusEnum","default":"active"},"id":{"type":"string","format":"uuid","title":"Id"},"category":{"anyOf":[{"$ref":"#/components/schemas/Category"},{"type":"null"}]},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"}},"type":"object","required":["name","price","id"],"title":"Item"},"ItemCreate":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"price":{"type":"number","title":"Price"},"category_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Category Id"},"image_urls":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Image Urls"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags"},"stock_quantity":{"type":"integer","title":"Stock Quantity","default":0},"low_stock_threshold":{"type":"integer","title":"Low Stock Threshold","default":10},"status":{"$ref":"#/components/schemas/ItemStatusEnum","default":"active"}},"type":"object","required":["name","price"],"title":"ItemCreate"},"ItemStatusEnum":{"type":"string","enum":["active","inactive","out_of_stock"],"title":"ItemStatusEnum"},"ItemUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price"},"category_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Category Id"},"image_urls":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Image Urls"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags"},"stock_quantity":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Stock Quantity"},"low_stock_threshold":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Low Stock Threshold"},"status":{"anyOf":[{"$ref":"#/components/schemas/ItemStatusEnum"},{"type":"null"}]}},"type":"object","title":"ItemUpdate"},"Order":{"properties":{"status":{"$ref":"#/components/schemas/OrderStatusEnum","default":"pending"},"subtotal":{"type":"number","title":"Subtotal"},"tax":{"type":"number","title":"Tax"},"points_discount":{"type":"number","title":"Points Discount"},"total":{"type":"number","title":"Total"},"payment_reference_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Payment Reference Id"},"id":{"type":"string","format":"uuid","title":"Id"},"user_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"invoice_file_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Invoice File Path"},"items":{"items":{"$ref":"#/components/schemas/OrderItem"},"type":"array","title":"Items","default":[]},"customer_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Customer Name"},"customer_phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Customer Phone"}},"type":"object","required":["subtotal","tax","points_discount","total","id","created_at"],"title":"Order"},"OrderCreate":{"properties":{"points_to_redeem":{"type":"integer","title":"Points To Redeem","default":0},"items":{"items":{"$ref":"#/components/schemas/OrderItemCreate"},"type":"array","title":"Items"},"customer_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Customer Name"},"customer_phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Customer Phone"}},"type":"object","required":["items"],"title":"OrderCreate"},"OrderItem":{"properties":{"item_id":{"type":"string","format":"uuid","title":"Item Id"},"quantity":{"type":"integer","title":"Quantity"},"id":{"type":"string","format":"uuid","title":"Id"},"order_id":{"type":"string","format":"uuid","title":"Order Id"},"item_name_snapshot":{"type":"string","title":"Item Name Snapshot"},"unit_price_at_purchase":{"type":"number","title":"Unit Price At Purchase"}},"type":"object","required":["item_id","quantity","id","order_id","item_name_snapshot","unit_price_at_purchase"],"title":"OrderItem"},"OrderItemCreate":{"properties":{"item_id":{"type":"string","format":"uuid","title":"Item Id"},"quantity":{"type":"integer","title":"Quantity"}},"type":"object","required":["item_id","quantity"],"title":"OrderItemCreate"},"OrderStatusEnum":{"type":"string","enum":["pending","paid","fulfilled","cancelled","refunded"],"title":"OrderStatusEnum"},"RoleEnum":{"type":"string","enum":["customer","admin"],"title":"RoleEnum"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"User":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"id":{"type":"string","format":"uuid","title":"Id"},"role":{"$ref":"#/components/schemas/RoleEnum"},"points_balance":{"type":"integer","title":"Points Balance"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"}},"type":"object","required":["email","id","role","points_balance","created_at"],"title":"User"},"UserCreate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"api/auth/login"}}}}}}